Move todo projection into control plane
